+++ src/main/java/org/betonquest/betonquest/compatibility/npc/citizens/CitizensNameWrapper.java
@@ -1,7 +1,7 @@
 package org.betonquest.betonquest.compatibility.npc.citizens;
 
-import net.citizensnpcs.api.CitizensAPI;
 import net.citizensnpcs.api.npc.NPC;
+import net.citizensnpcs.api.npc.NPCRegistry;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.npc.Npc;
@@ -14,6 +14,11 @@ import org.jetbrains.annotations.Nullable;
  */
 public class CitizensNameWrapper implements NpcWrapper<NPC> {
     /**
+     * Source Registry of NPCs to use.
+     */
+    private final NPCRegistry registry;
+
+    /**
      * Name of the Npc.
      */
     private final Variable<String> npcName;
@@ -21,9 +26,11 @@ public class CitizensNameWrapper implements NpcWrapper<NPC> {
     /**
      * Create a new Citizens Npc Wrapper.
      *
-     * @param npcName the name of the Npc
+     * @param registry the registry of NPCs to use
+     * @param npcName  the name of the Npc
      */
-    public CitizensNameWrapper(final Variable<String> npcName) {
+    public CitizensNameWrapper(final NPCRegistry registry, final Variable<String> npcName) {
+        this.registry = registry;
         this.npcName = npcName;
     }
 
@@ -31,7 +38,7 @@ public class CitizensNameWrapper implements NpcWrapper<NPC> {
     public Npc<NPC> getNpc(@Nullable final Profile profile) throws QuestException {
         NPC selectedNpc = null;
         final String npcName = this.npcName.getValue(profile);
-        for (final NPC npc : CitizensAPI.getNPCRegistry()) {
+        for (final NPC npc : registry) {
             if (npc.getName().equals(npcName)) {
                 if (selectedNpc == null) {
                     selectedNpc = npc;