+++ src/main/java/org/betonquest/betonquest/compatibility/npc/citizens/CitizensWrapper.java
@@ -1,7 +1,7 @@
 package org.betonquest.betonquest.compatibility.npc.citizens;
 
-import net.citizensnpcs.api.CitizensAPI;
 import net.citizensnpcs.api.npc.NPC;
+import net.citizensnpcs.api.npc.NPCRegistry;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.npc.Npc;
@@ -14,6 +14,11 @@ import org.jetbrains.annotations.Nullable;
  */
 class CitizensWrapper implements NpcWrapper<NPC> {
     /**
+     * Source Registry of NPCs to use.
+     */
+    private final NPCRegistry registry;
+
+    /**
      * Id of the Npc.
      */
     private final Variable<Number> npcId;
@@ -21,16 +26,18 @@ class CitizensWrapper implements NpcWrapper<NPC> {
     /**
      * Create a new Citizens Npc Wrapper.
      *
-     * @param npcId the id of the Npc, greater or equals to zero
+     * @param registry the registry of NPCs to use
+     * @param npcId    the id of the Npc, greater or equals to zero
      */
-    public CitizensWrapper(final Variable<Number> npcId) {
+    public CitizensWrapper(final NPCRegistry registry, final Variable<Number> npcId) {
+        this.registry = registry;
         this.npcId = npcId;
     }
 
     @Override
     public Npc<NPC> getNpc(@Nullable final Profile profile) throws QuestException {
         final int npcId = this.npcId.getValue(profile).intValue();
-        final NPC npc = CitizensAPI.getNPCRegistry().getById(npcId);
+        final NPC npc = registry.getById(npcId);
         if (npc == null) {
             throw new QuestException("NPC with ID " + npcId + " not found");
         }