+++ CHANGELOG.md
@@ -40,6 +40,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - menu `slots` and bound `item` now support variables, `menu item` now supports variable `item`
 - `menu` event now supports the `update` action to update the currently opened menu
 - `menu` conversation IO settings to configure the scroll down and up buttons
+- `citizens` npcs are now used only from default registry
 ### Changed
 - Spigot is no longer supported, paper is now required 
 - message.yml file was deleted and instead the lang folder now contains all translations